Patient Management
For now, a hospital personnel is able to view available patients in the system, and thereby be able to view their details, and the list of files and records they have. The next step is going to be to allow them to view files inside the record and the ability to open or close the record.