File added
So far, I have managed to record files with and without attachments, I managed to make sure that I can display the files inside a record, and I am also capable of opening and closing records. Now, the next thing is getting the ability to view a file details by clicking on it.