Update logo & certificate
Today, I managed to make the system capable of uploading a new logo, or update an existing one as it can do it with the certificate. I also made it possible to create a file which will at the end have a logo for the hospital from which it was created from.